+++ pkg/models/configs.go
@@ -15,6 +15,6 @@ type Redis struct {
 }
 
 type PercentageDivision struct {
-	Api1Percent int
-	Api2Percent int
+	Api1Percent float64
+	Api2Percent float64
 }
